[GroheOndus] Add support for GROHE SENSE device
With this change, the binding now also supports the slightly different
device GROHE SENSE, which does not have the same channels as the SENSE
GUARD and provides some other information.

Fixes #3898
