+++ addons/binding/org.openhab.binding.groheondus/.classpath
@@ -3,6 +3,6 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src/main/java"/>
-	<classpathentry kind="lib" path="lib/ondus-api-0.0.2.jar"/>
+	<classpathentry kind="lib" path="lib/ondus-api-0.0.3.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
