+++ addons/binding/org.openhab.binding.groheondus/build.properties
@@ -5,4 +5,4 @@ bin.includes = META-INF/,\
                OSGI-INF/,\
                ESH-INF/,\
                about.html,\
-               lib/ondus-api-0.0.2.jar
+               lib/ondus-api-0.0.3.jar
