+++ addons/binding/org.openhab.binding.groheondus/src/main/java/org/openhab/binding/groheondus/internal/discovery/GroheOndusDiscoveryService.java
@@ -8,7 +8,7 @@
  */
 package org.openhab.binding.groheondus.internal.discovery;
 
-import static org.openhab.binding.groheondus.internal.GroheOndusBindingConstants.THING_TYPE_APPLIANCE;
+import static org.openhab.binding.groheondus.internal.GroheOndusBindingConstants.*;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -23,9 +23,11 @@ import org.eclipse.smarthome.config.discovery.DiscoveryResult;
 import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.grohe.ondus.api.OndusService;
-import org.grohe.ondus.api.model.Appliance;
+import org.grohe.ondus.api.model.BaseAppliance;
 import org.grohe.ondus.api.model.Location;
 import org.grohe.ondus.api.model.Room;
+import org.grohe.ondus.api.model.SenseAppliance;
+import org.grohe.ondus.api.model.SenseGuardAppliance;
 import org.openhab.binding.groheondus.internal.handler.GroheOndusAccountHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -46,7 +48,7 @@ public class GroheOndusDiscoveryService extends AbstractDiscoveryService {
     private final GroheOndusAccountHandler bridgeHandler;
 
     public GroheOndusDiscoveryService(GroheOndusAccountHandler bridgeHandler) {
-        super(Collections.singleton(THING_TYPE_APPLIANCE), 30);
+        super(Collections.singleton(THING_TYPE_SENSEGUARD), 30);
         logger.debug("initialize discovery service");
         this.bridgeHandler = bridgeHandler;
         this.activate(null);
@@ -61,7 +63,7 @@ public class GroheOndusDiscoveryService extends AbstractDiscoveryService {
             logger.debug("No instance of OndusService given.", e);
             return;
         }
-        List<Appliance> discoveredAppliances = new ArrayList<>();
+        List<BaseAppliance> discoveredAppliances = new ArrayList<>();
         try {
             discoveredAppliances = getApplianceList(service);
         } catch (IOException e) {
@@ -71,7 +73,17 @@ public class GroheOndusDiscoveryService extends AbstractDiscoveryService {
 
         discoveredAppliances.forEach(appliance -> {
             ThingUID bridgeUID = bridgeHandler.getThing().getUID();
-            ThingUID thingUID = new ThingUID(THING_TYPE_APPLIANCE, appliance.getApplianceId());
+            ThingUID thingUID = null;
+            switch (appliance.getType()) {
+                case SenseGuardAppliance.TYPE:
+                    thingUID = new ThingUID(THING_TYPE_SENSEGUARD, appliance.getApplianceId());
+                    break;
+                case SenseAppliance.TYPE:
+                    thingUID = new ThingUID(THING_TYPE_SENSE, appliance.getApplianceId());
+                    break;
+                default:
+                    return;
+            }
 
             Map<String, Object> properties = new HashMap<>();
             properties.put(PROPERTY_LOCATION_ID, appliance.getRoom().getLocation().getId());
@@ -86,8 +98,8 @@ public class GroheOndusDiscoveryService extends AbstractDiscoveryService {
         });
     }
 
-    private List<Appliance> getApplianceList(OndusService service) throws IOException {
-        List<Appliance> discoveredAppliances = new ArrayList<>();
+    private List<BaseAppliance> getApplianceList(OndusService service) throws IOException {
+        List<BaseAppliance> discoveredAppliances = new ArrayList<>();
 
         for (Location location : service.getLocations()) {
             for (Room room : service.getRooms(location)) {
