+++ addons/binding/org.openhab.binding.groheondus/src/main/java/org/openhab/binding/groheondus/internal/handler/GroheOndusAccountHandler.java
@@ -70,7 +70,7 @@ public class GroheOndusAccountHandler extends BaseBridgeHandler {
             updateStatus(ThingStatus.ONLINE);
 
             scheduler.submit(() -> getThing().getThings().forEach(thing -> {
-                GroheOndusApplianceHandler thingHandler = (GroheOndusApplianceHandler) thing.getHandler();
+                GroheOndusSenseGuardHandler thingHandler = (GroheOndusSenseGuardHandler) thing.getHandler();
                 if (thingHandler != null) {
                     thingHandler.updateChannels();
                 }
