+++ addons/binding/org.openhab.binding.groheondus/src/main/java/org/openhab/binding/groheondus/internal/handler/GroheOndusHandlerFactory.java
@@ -39,8 +39,8 @@ public class GroheOndusHandlerFactory extends BaseThingHandlerFactory {
 
     private final Map<ThingUID, @Nullable ServiceRegistration<?>> discoveryServiceRegs = new HashMap<>();
 
-    private static final Collection<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = Arrays.asList(THING_TYPE_APPLIANCE,
-            THING_TYPE_BRIDGE_ACCOUNT);
+    private static final Collection<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = Arrays.asList(THING_TYPE_SENSEGUARD,
+            THING_TYPE_SENSE, THING_TYPE_BRIDGE_ACCOUNT);
 
     @Override
     public boolean supportsThingType(ThingTypeUID thingTypeUID) {
@@ -55,8 +55,10 @@ public class GroheOndusHandlerFactory extends BaseThingHandlerFactory {
             GroheOndusAccountHandler handler = new GroheOndusAccountHandler((Bridge) thing);
             registerDeviceDiscoveryService(handler);
             return handler;
-        } else if (THING_TYPE_APPLIANCE.equals(thingTypeUID)) {
-            return new GroheOndusApplianceHandler(thing);
+        } else if (THING_TYPE_SENSEGUARD.equals(thingTypeUID)) {
+            return new GroheOndusSenseGuardHandler(thing);
+        } else if (THING_TYPE_SENSE.equals(thingTypeUID)) {
+            return new GroheOndusSenseHandler(thing);
         }
 
         return null;
